+++ .gitignore
@@ -1,0 +1,25 @@
+.haikro-cache/
+node-linux-x64
+/node_modules/
+/bower_components/
+/__attic/
+/.sass-cache/
+/.node-version
+.DS_Store
+npm-debug.log
+/node-v*
+/tmp/
+*.env*
+.idea/
+/test/browser/reports
+/test/browser/screenshots
+/test/visual/screenshots
+/pa11y_screenCapture/*
+/npm-debug.log.*
+/public*
+/pa11y_screenCapture/
+package-lock.json
+.editorconfig
+.eslintrc.js
+.pa11yci.js
+.stylelintrc
