+++ lib/swg-client.js
@@ -9,6 +9,12 @@ module.exports = () => {
   script.src = 'https://subscribe.sandbox.google.com/swglib/swg.js';
   script.async = true;
   script.id = 'swg-client';
+
+  // If we need to manually init, we need to let the swg client know.
+  if (document.querySelectorAll('[data-n-swg-manual-init]').length) {
+    script.setAttribute('subscriptions-control', 'manual');
+  }
+  
   hasClientAlready = true;
   
   document.body.appendChild(script);
