+++ main.scss
@@ -1,1 +1,1 @@
-// SwG styles will go in here.
+@import 'lib/subscribe-button/styles';
