+++ views/button.html
@@ -1,3 +1,3 @@
-<a class="o-subs-card__select-button barrier__button-primary--big" data-trackable="swg-{{appName}}" data-n-swg-button n-swg-button-sku="{{sku}}">
+<a class="o-buttons o-buttons--primary o-buttons--big swg__button" data-trackable="swg-{{appName}}" data-n-swg-button data-n-swg-button-sku="{{sku}}" data-n-swg-manual-init="{{manualInit}}">
   Subscribe with Google
 </a>
